+++ packages/core/src/Table/TableFoot.js
@@ -8,8 +8,8 @@ import React from 'react'
  * @example import { TableFoot } from '@dhis2/ui-core'
  * @see Live demo: {@link /demo/?path=/story/table--static-layout|Storybook}
  */
-export const TableFoot = ({ children, className, dataTest }) => (
-    <tfoot className={className} data-test={dataTest}>
+export const TableFoot = ({ children, className, dataTest, role }) => (
+    <tfoot className={className} data-test={dataTest} role={role}>
         {children}
     </tfoot>
 )
@@ -23,6 +23,7 @@ TableFoot.defaultProps = {
  * @static
  * @prop {TableRow|Array.<TableRow>} [children]
  * @prop {string} [className]
+ * @prop {string} [role]
  * @prop {string} [dataTest]
  */
 TableFoot.propTypes = {
@@ -30,4 +31,5 @@ TableFoot.propTypes = {
     children: PropTypes.node,
     className: PropTypes.string,
     dataTest: PropTypes.string,
+    role: PropTypes.string,
 }