+++ packages/core/src/Table/TableHead.js
@@ -8,8 +8,8 @@ import React from 'react'
  * @example import { TableHead } from '@dhis2/ui-core'
  * @see Live demo: {@link /demo/?path=/story/table--static-layout|Storybook}
  */
-export const TableHead = ({ children, className, dataTest }) => (
-    <thead className={className} data-test={dataTest}>
+export const TableHead = ({ children, className, dataTest, role }) => (
+    <thead className={className} data-test={dataTest} role={role}>
         {children}
     </thead>
 )
@@ -23,6 +23,7 @@ TableHead.defaultProps = {
  * @static
  * @prop {TableRowHead|Array.<TableRowHead>} [children]
  * @prop {string} [className]
+ * @prop {string} [role]
  * @prop {string} [dataTest]
  */
 TableHead.propTypes = {
@@ -30,4 +31,5 @@ TableHead.propTypes = {
     children: PropTypes.node,
     className: PropTypes.string,
     dataTest: PropTypes.string,
+    role: PropTypes.string,
 }